fix: handle pre-migration state — cast discount TEXT→INTEGER, catch missing tables
- Add ::INTEGER cast to discount COALESCE in students.ts, payments.ts, individual-payments.ts
  (prevents PostgreSQL type mismatch when discount is still TEXT before migration)
- Wrap individual_balances query in try-catch in payments/overview route
  (table may not exist before migration is run)

+++ src/lib/individual-payments.ts
@@ -119,7 +119,7 @@ export async function calculateIndividualAmount(
 ): Promise<{ lesson_price: number; discount_percent: number; effective_price: number; total: number }> {
   const lessonPrice = await getLessonPrice();
   const student = await get<{ discount: number | null }>(
-    `SELECT COALESCE(discount, 0) as discount FROM students WHERE id = $1`,
+    `SELECT COALESCE(discount::INTEGER, 0) as discount FROM students WHERE id = $1`,
     [studentId]
   );
   const discountPercent = student?.discount || 0;
